pyproject.toml
[tool.poetry]
name = "its-tls-app"
version = "0.1.0"
description = "A simple web app for viewing and monitoring customer certificates"
authors = ["Mason Putney"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
jinja2 = "^3.1.2"
datetime = "^5.2"
requests = "^2.31.0"
python-multipart = "^0.0.6"
sqlalchemy = "^2.0.21"
python-dateutil = "^2.8.2"
uvicorn = "^0.23.2"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"
msal = "^1.24.1"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
